requirements.txt
absl-py==0.8.0
astor==0.8.0
cycler==0.10.0
Cython==0.29.14
decorator==4.4.1
gast==0.3.2
grpcio==1.23.0
h5py==2.10.0
imageio==2.6.1
imgaug==0.3.0
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Markdown==3.1.1
matplotlib==3.1.1
networkx==2.4
numpy==1.17.2
opencv-python==4.1.1.26
opencv-python-headless==4.1.2.30
pandas==0.25.3
Pillow==6.1.0
progressbar2==3.47.0
protobuf==3.9.1
pyparsing==2.4.2
python-dateutil==2.8.0
python-utils==2.3.0
pytz==2019.3
PyWavelets==1.1.1
PyYAML==5.2
scikit-image==0.16.2
scipy==1.3.3
seaborn==0.9.0
Shapely==1.6.4.post2
six==1.12.0
tensorboard==1.12.2
tensorflow-gpu==1.12.0
termcolor==1.1.0
Werkzeug==0.15.6
